Move DrawVisualiser to own namespace and split out classes.
